vps-agent/requirements.txt
fastapi
uvicorn[standard]
psutil
docker
python-dotenv
httpx
